Rearranging the migration runner
[skip ci]
